+++ changelog.md
@@ -19,6 +19,8 @@
 -   Allow warn btn js to work on ajaxed elements
 -   Remove blank sortable column on overview screen if model is not sortable
 -   Add "canDelete" param to overview screen to hide the delete button
+-   Add "Save and continue" button on form screen
+-   Add "Model created" flash message as well as updated
 
 
 # Version 1.0.4 (2015-06-15)